fixed skill average roles and refactored to function
