+++ src/lib/hypixelSkillCalc.ts
@@ -1,3 +1,5 @@
+import { profileMember } from "../typings/Hypixel";
+
 export const basedskillRequiredExp: { [key: number]: number } = {
   1: 50,
   2: 125,
@@ -75,6 +77,46 @@ export const skillLevelCalculator = (
 
   return xp > 0 ? skillLevelCalculator(xp, level + 1, type) : level;
 };
+
+export const skillAverageCalculator = (profile: profileMember) => {
+  let skill_alchemy = skillLevelCalculator(
+    profile?.experience_skill_alchemy
+  );
+  let skill_mining = skillLevelCalculator(
+    profile?.experience_skill_mining
+  );
+  let skill_enchanting = skillLevelCalculator(
+    profile?.experience_skill_enchanting
+  );
+  let skill_farming = skillLevelCalculator(
+    profile?.experience_skill_farming
+  );
+  let skill_fishing = skillLevelCalculator(
+    profile?.experience_skill_fishing
+  );
+  let skill_foraging = skillLevelCalculator(
+    profile?.experience_skill_foraging
+  );
+  let skill_taming = skillLevelCalculator(
+    profile?.experience_skill_taming
+  );
+  let skill_combat = skillLevelCalculator(
+    profile?.experience_skill_combat
+  );
+
+  let average =
+    (skill_alchemy +
+      skill_mining +
+      skill_enchanting +
+      skill_farming +
+      skill_fishing +
+      skill_foraging +
+      skill_taming +
+      skill_combat) /
+    8;
+
+    return average;
+}
 
 export const dungeonSkillRequiredExp: { [key: number]: number } = {
   1: 50,
